refactoring: make common line template
