+++ packages/dx-react-chart-material-ui/src/templates/axis/line.jsx
@@ -1,4 +1,3 @@
-
 import { Axis } from '@devexpress/dx-react-chart';
 import { withClassName, getBorderColor } from '../utils';
