+++ packages/dx-react-chart/src/templates/axis/line.jsx
@@ -3,10 +3,12 @@ import * as PropTypes from 'prop-types';
 
 export class Line extends React.PureComponent {
   render() {
-    const { width, height, ...restProps } = this.props;
+    const {
+      x1, x2, y1, y2, ...restProps
+    } = this.props;
     return (
       <path
-        d={`M 0 0 L ${width} ${height}`}
+        d={`M ${x1} ${y1} L ${x2} ${y2}`}
         {...restProps}
       />
     );
@@ -14,6 +16,8 @@ export class Line extends React.PureComponent {
 }
 
 Line.propTypes = {
-  width: PropTypes.number.isRequired,
-  height: PropTypes.number.isRequired,
+  x1: PropTypes.number.isRequired,
+  x2: PropTypes.number.isRequired,
+  y1: PropTypes.number.isRequired,
+  y2: PropTypes.number.isRequired,
 };
