+++ packages/dx-react-chart/src/templates/axis/line.test.jsx
@@ -2,21 +2,28 @@ import * as React from 'react';
 import { shallow } from 'enzyme';
 import { Line } from './line';
 
-describe('Line', () => {
-  const defaultProps = { width: 80, height: 70 };
+const defaultProps = {
+  x1: 1, x2: 2, y1: 3, y2: 4,
+};
 
-  it('should render line', () => {
+describe('Line', () => {
+  it('should render line with correct coordinates', () => {
     const { d } = shallow((
       <Line {...defaultProps} />
     )).find('path').props();
 
-    expect(d).toBe('M 0 0 L 80 70');
+    expect(d).toBe('M 1 3 L 2 4');
   });
 
   it('should pass the rest property to the root element', () => {
     const tree = shallow(<Line {...defaultProps} customProperty />);
     const { customProperty } = tree.find('path').props();
-
     expect(customProperty).toBeTruthy();
+  });
+
+  it('should pass the className to the root element', () => {
+    const tree = shallow(<Line {...defaultProps} className />);
+    const { className } = tree.find('path').props();
+    expect(className).toBeTruthy();
   });
 });
